stateを子コンポーネントで管理するようにしてみた
