+++ src/components/Furigana.jsx
@@ -1,7 +1,9 @@
-import React from 'react'
+import React, { useState } from 'react'
 
 export const Furigana = (props) => {
-  const {furigana, onChangeFurigana} = props;
+  // const {furigana, onChangeFurigana} = props;
+  const [furigana, setFurigana] = useState("");
+  const onChangeFurigana = (event) => setFurigana(event.target.value);
 
   const inputFurigana = furigana !== '' ? {
     fontSize: "1rem",
